wire: Wire v2 — 32-bit epochs and generations, 48-bit crypto counters (#29, #48)
Wire v1 carried 16-bit link/end epochs and firmware consumed one per boot
(deep-sleep wake included). After 65,535 boots the epoch wrapped: every
peer's replay floor and route table rejected the node forever, and its own
TX counter lease refused the "older" epoch, so it could not send either.

Wire v2 (major 2) widens the epochs to 32 bits and narrows the crypto
counters to 48 bits, so the header stays 88 bytes and the 128-byte payload
still fits the 250-byte ESP-NOW body:

- header: link epoch u32 @68, end epoch u32 @72, link counter u48 @76,
  end counter u48 @82; counters above kMaxCryptoCounter are refused and a
  lease exhausts there instead of wrapping (a new epoch is required).
- AEAD nonce (dev PSK): scope | direction | epoch u32 | counter u48;
  key derivation binds the 32-bit epoch.
- ROUTE_UPDATE records carry a u32 origin generation (16 B, 7 per frame).
- APPLIED execution lease: session u32 | end epoch u32 | incarnation u64
  (still 16 B; the nonzero session replaces the magic).
- persisted TX counter records (32 B) and replay floor records gain a
  layout byte (2); v1 records fail closed with IntegrityError — devices
  flashed with v1 firmware need an NVS erase.
- C ABI: RL_ABI_VERSION 2, epoch/generation fields are uint32_t.
- firmware derives epochs and route generation straight from the 32-bit
  persisted boot session (no % 0xFFFF fold).
- Rust wire codec, test cipher and golden generator follow; shared
  golden vectors regenerated, plus an invalid wire_v1_frame vector.
- protocol/semantics.json end_aad_fields: end epoch 4 B, end counter 6 B.
- tests: fixtures that hand-build v1 route records / header offsets move
  to v2 (the tombstone fill loop is now bounded so a rejected format
  fails instead of hanging); new regressions walk a peer pair across the
  old 65,535 wrap point (counter lease + replay floor), check route
  generations past the u16 budget, and run reliable delivery between
  nodes with epochs > 70,000.
- docs: wire-protocol (v2 change table), compatibility, STATUS,
  implementation README, crash-time-resources, protocol READMEs.

+++ components/routeloom/include/routeloom/node.hpp
@@ -20,12 +20,12 @@ struct NodeConfig {
   NetworkId network{0};
   NodeId node{kInvalidNodeId};
   std::uint32_t message_session{0};
-  std::uint16_t link_epoch{1};
-  std::uint16_t end_epoch{1};
+  std::uint32_t link_epoch{1};
+  std::uint32_t end_epoch{1};
   // Origin generation for this node's own route source. Must be persisted
   // monotonic and incremented on every boot; a restarted node advertises a
   // higher generation so peers discard its previous-incarnation route state.
-  std::uint16_t route_generation{1};
+  std::uint32_t route_generation{1};
   std::uint32_t route_advertisement_period_ms{5000};
   std::uint32_t route_lifetime_ms{15000};
   // §14 management airtime budget gate (03-congestion.md §8, radio.md
